requirements.txt
Flask[async]
Flask-SQLAlchemy
Flask-Login
Flask-Migrate
Flask-CORS
Werkzeug
gunicorn
pytz
azure-ai-formrecognizer
python-dotenv
pandas
xlsxwriter
psycopg2-binary
dash
dash-core-components
dash-html-components
dash-bootstrap-components
plotly
flask_wtf
msal
requests
